Move `OutlineDecoration` to devtools_app_shared
