+++ packages/devtools_app/lib/src/screens/memory/panes/diff/widgets/class_details/class_details.dart
@@ -5,7 +5,6 @@
 import 'package:devtools_app_shared/ui.dart';
 import 'package:flutter/widgets.dart';
 
-import '../../../../../../shared/common_widgets.dart';
 import '../../../../shared/heap/heap.dart';
 import '../../controller/class_data.dart';
 import 'path.dart';
