+++ packages/devtools_app/lib/src/screens/network/network_request_inspector.dart
@@ -6,7 +6,6 @@ import 'package:devtools_app_shared/ui.dart';
 import 'package:flutter/material.dart';
 
 import '../../shared/analytics/constants.dart' as gac;
-import '../../shared/common_widgets.dart';
 import '../../shared/http/http_request_data.dart';
 import '../../shared/ui/tab.dart';
 import 'network_controller.dart';
