+++ packages/devtools_app/lib/src/shared/common_widgets.dart
@@ -943,73 +943,6 @@ class OutlinedRowGroup extends StatelessWidget {
   }
 }
 
-class OutlineDecoration extends StatelessWidget {
-  const OutlineDecoration({
-    Key? key,
-    this.child,
-    this.showTop = true,
-    this.showBottom = true,
-    this.showLeft = true,
-    this.showRight = true,
-  }) : super(key: key);
-
-  factory OutlineDecoration.onlyBottom({required Widget? child}) =>
-      OutlineDecoration(
-        showTop: false,
-        showLeft: false,
-        showRight: false,
-        child: child,
-      );
-
-  factory OutlineDecoration.onlyTop({required Widget? child}) =>
-      OutlineDecoration(
-        showBottom: false,
-        showLeft: false,
-        showRight: false,
-        child: child,
-      );
-
-  factory OutlineDecoration.onlyLeft({required Widget? child}) =>
-      OutlineDecoration(
-        showBottom: false,
-        showTop: false,
-        showRight: false,
-        child: child,
-      );
-
-  factory OutlineDecoration.onlyRight({required Widget? child}) =>
-      OutlineDecoration(
-        showBottom: false,
-        showTop: false,
-        showLeft: false,
-        child: child,
-      );
-
-  final bool showTop;
-  final bool showBottom;
-  final bool showLeft;
-  final bool showRight;
-
-  final Widget? child;
-
-  @override
-  Widget build(BuildContext context) {
-    final color = Theme.of(context).focusColor;
-    final border = BorderSide(color: color);
-    return Container(
-      decoration: BoxDecoration(
-        border: Border(
-          left: showLeft ? border : BorderSide.none,
-          right: showRight ? border : BorderSide.none,
-          top: showTop ? border : BorderSide.none,
-          bottom: showBottom ? border : BorderSide.none,
-        ),
-      ),
-      child: child,
-    );
-  }
-}
-
 class ThickDivider extends StatelessWidget {
   const ThickDivider({super.key});
